coalesce scattered dupes of unit-test utilities
