+++ tests/QTestExtensions.h
@@ -37,6 +37,10 @@
 //
 
 
+float getErrorDifference(const float& a, const float& b) {
+    return fabsf(a - b);
+}
+
 // Generates a QCOMPARE-style failure message that can be passed to QTest::qFail.
 //
 // Formatting looks like this:
@@ -281,3 +285,20 @@ bool compareData (const char* data, const char* expectedData, size_t length) {
 
 #define COMPARE_DATA(actual, expected, length) \
     QCOMPARE_WITH_EXPR((ByteData ( actual, length )), (ByteData ( expected, length )), compareData(actual, expected, length))
+
+
+// Produces a relative error test for float usable QCOMPARE_WITH_LAMBDA.
+inline auto errorTest (float actual, float expected, float acceptableRelativeError)
+-> std::function<bool ()> {
+    return [actual, expected, acceptableRelativeError] () {
+        if (fabsf(expected) <= acceptableRelativeError) {
+            return fabsf(actual - expected) < fabsf(acceptableRelativeError);
+        }
+        return fabsf((actual - expected) / expected) < fabsf(acceptableRelativeError);
+    };
+}
+
+#define QCOMPARE_WITH_RELATIVE_ERROR(actual, expected, relativeError) \
+    QCOMPARE_WITH_LAMBDA(actual, expected, errorTest(actual, expected, relativeError))
+
+
